requirements.txt
numpy
pandas
scikit-learn
nltk
pdfplumber
python-docx
streamlit
